+++ src/main/java/mongkey/maeilmail/domain/Template.java
@@ -13,7 +13,8 @@ import mongkey.maeilmail.common.domain.BaseTimeEntity;
 public class Template extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long template_id;
+    @Column(name = "template_id")
+    private Long id;
 
     @Column(length = 255, nullable = false)
     private String title;
@@ -29,8 +30,8 @@ public class Template extends BaseTimeEntity {
     private Admin admin;
 
     @Builder
-    public Template(Long id, Long admin_id, String title, String content, Integer copy_count){
-        this.admin_id = admin_id;
+    public Template(Long id,  String title, String content, Integer copy_count, Admin admin){
+        this.admin = admin;
         this.title = title;
         this.content = content;
         this.copy_count = copy_count;
@@ -44,13 +45,13 @@ public class Template extends BaseTimeEntity {
         this.id = id;
     }
 
-    public Long getAdmin_id() {
-        return admin_id;
-    }
-
-    public void setAdmin_id(Long admin_id) {
-        this.admin_id = admin_id;
-    }
+//    public Long getAdmin_id() {
+//        return admin_id;
+//    }
+//
+//    public void setAdmin_id(Long admin_id) {
+//        this.admin_id = admin_id;
+//    }
 
     public String getTitle() {
         return title;
