+++ src/main/java/mongkey/maeilmail/dto/post/request/SavePostRequestDto.java
@@ -1,4 +1,4 @@
-package mongkey.maeilmail.dto.post;
+package mongkey.maeilmail.dto.post.request;
 
 import lombok.Getter;
 import mongkey.maeilmail.domain.Post;
@@ -6,11 +6,12 @@ import mongkey.maeilmail.domain.User;
 import mongkey.maeilmail.domain.enums.CategoryType;
 
 @Getter
-public class UpdatePostRequestDto {
+public class SavePostRequestDto {
     private Long user_id;
     private CategoryType category;
     private String title;
     private String content;
+
 //        private PostType postType;
 
     //SavePostRequestDto를 실제 엔티티로 변환
@@ -26,7 +27,7 @@ public class UpdatePostRequestDto {
     @Override
     public String toString() {
         return "SavePostRequestDto{" +
-                "user_id='" + user_id+ '\'' +
+                "user_id='" + user_id + '\'' +
                 ", category=" + category +
                 ", title='" + title + '\'' +
                 ", content='" + content + '\'' +
