+++ src/main/java/mongkey/maeilmail/dto/post/request/UpdatePostRequestDto.java
@@ -1,4 +1,4 @@
-package mongkey.maeilmail.dto.post;
+package mongkey.maeilmail.dto.post.request;
 
 import lombok.Getter;
 import mongkey.maeilmail.domain.Post;
@@ -6,12 +6,11 @@ import mongkey.maeilmail.domain.User;
 import mongkey.maeilmail.domain.enums.CategoryType;
 
 @Getter
-public class SavePostRequestDto {
+public class UpdatePostRequestDto {
     private Long user_id;
     private CategoryType category;
     private String title;
     private String content;
-
 //        private PostType postType;
 
     //SavePostRequestDto를 실제 엔티티로 변환
@@ -27,7 +26,7 @@ public class SavePostRequestDto {
     @Override
     public String toString() {
         return "SavePostRequestDto{" +
-                "user_id='" + user_id + '\'' +
+                "user_id='" + user_id+ '\'' +
                 ", category=" + category +
                 ", title='" + title + '\'' +
                 ", content='" + content + '\'' +
