+++ src/main/java/mongkey/maeilmail/service/MemberService.java
@@ -6,11 +6,9 @@ import mongkey.maeilmail.common.response.ApiResponse;
 import mongkey.maeilmail.common.response.Error;
 import mongkey.maeilmail.common.response.Success;
 import mongkey.maeilmail.domain.Admin;
-import mongkey.maeilmail.domain.Post;
 import mongkey.maeilmail.domain.User;
 import mongkey.maeilmail.dto.admin.request.JoinAdminRequestDto;
 import mongkey.maeilmail.dto.admin.request.LoginAdminRequestDto;
-import mongkey.maeilmail.dto.post.SavePostRequestDto;
 import mongkey.maeilmail.dto.user.request.JoinUserRequestDto;
 import mongkey.maeilmail.dto.user.request.LoginUserRequestDto;
 import mongkey.maeilmail.repository.AdminRepository;
